Added inverse routine
