+++ examples/reverse_hello_keyedge/clean.sh
@@ -1,0 +1,14 @@
+rm eapp/eapp
+rm eapp/eapp.o
+rm host/host
+rm host/host.o
+rm keyedge/flatbuffers_common_builder.h
+rm keyedge/flatbuffers_common_reader.h
+rm keyedge/ocalls.fbs
+rm keyedge/ocalls_builder.h
+rm keyedge/ocalls_common.h
+rm keyedge/ocalls_eapp.h
+rm keyedge/ocalls_host.h
+rm keyedge/ocalls_reader.h
+rm keyedge/ocalls_verifier.h
+
