+++ examples/reverse_hello_keyedge/host/host.cpp
@@ -1,0 +1,29 @@
+#include "ocalls_host.h"
+
+Keystone enclave;
+
+int main(int argc, char** argv) {
+	Params params;
+
+	params.setFreeMemSize(2 * 1024 * 1024);
+	params.setUntrustedMem(DEFAULT_UNTRUSTED_PTR, 2 * 1024 * 1024);
+
+	enclave.init(argv[1], argv[2], params);
+
+	enclave.registerOcallDispatch(incoming_call_dispatch);
+	
+	printf("Registering...\n");
+	fflush(stdout);
+
+	edge_call_init_internals((uintptr_t)enclave.getSharedBuffer(),
+		enclave.getSharedBufferSize());
+	
+	printf("Running the enclave...\n");
+	fflush(stdout);
+
+	enclave.run();
+	printf ("%s\n", get_msg().content);
+
+	return 0;
+}
+
