+++ include/emitter/emit_callee.h
@@ -1,0 +1,2 @@
+std::string emit_function_callee(std::shared_ptr<function_information> f, size_t indent = 0); 
+
