+++ include/emitter/emit_caller.h
@@ -1,0 +1,2 @@
+std::string emit_function_caller(std::shared_ptr<function_information> f, size_t indent = 0); 
+
