--- include/emitter/emit_ihost.h
@@ -1,2 +1,0 @@
-std::string emit_ihost(size_t indent = 0);
-
